# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Odace Data Pipeline API",
    "description": "\n## Odace Data Pipeline Platform\n\nA cloud-based data pipeline platform for managing data ingestion, transformation, and processing.\n\n### Features\n\n* **Multi-layer architecture**: Bronze (ingestion), Silver (transformation), Gold (aggregation)\n* **Delta Lake storage**: Efficient data lake with ACID transactions\n* **API-driven**: RESTful API for pipeline orchestration\n* **Authenticated**: Secure API key authentication\n* **Scalable**: Runs on Google Cloud Run\n\n### Authentication\n\nAll endpoints require an API key in the Authorization header:\n\n```\nAuthorization: Bearer sk_live_YOUR_API_KEY\n```\n\n### Available Pipelines\n\n**Bronze Layer (Data Ingestion)**:\n- accueillants, geo, logement, gares, lignes, zones_attraction\n\n**Silver Layer (Data Transformation)**:\n- accueillants, geo, logement, gares, lignes, zones_attraction\n\n### Getting Started\n\n1. Obtain an API key from your administrator\n2. Include the API key in all requests: `Authorization: Bearer sk_live_...`\n3. Use `/api/pipeline/list` to see available pipelines\n4. Run pipelines with `/api/bronze/{pipeline_name}` or `/api/silver/{pipeline_name}`\n\nFor detailed documentation, visit the project repository.\n    ",
    "contact": {
      "name": "Odace Data Team",
      "email": "data@odace.example.com"
    },
    "license": {
      "name": "Internal Use Only"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/files/upload": {
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Upload File",
        "description": "Upload a file to the raw data folder in GCS.\n\nArgs:\n    domain: Domain/folder name (e.g., 'accueillants', 'geo', 'logement')\n    file: File to upload\n    api_key: API key for authentication\n    \nReturns:\n    File upload response with destination path",
        "operationId": "upload_file_api_files_upload_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_files_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/list": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "List Files",
        "description": "List files in raw data storage.\n\nArgs:\n    domain: Optional domain filter\n    api_key: API key for authentication\n    \nReturns:\n    List of files",
        "operationId": "list_files_api_files_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/catalog": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Catalog",
        "description": "Get the complete data catalog with all schemas and tables.\n\nScans GCS for Delta tables in bronze, silver, and gold layers.",
        "operationId": "get_catalog_api_data_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/data/table/{layer}/{table}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Table Metadata",
        "description": "Get metadata for a specific table including schema and row count.\n\nArgs:\n    layer: Layer name (bronze, silver, gold)\n    table: Table name",
        "operationId": "get_table_metadata_api_data_table__layer___table__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer"
            }
          },
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/preview/{layer}/{table}": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Preview Table",
        "description": "Get a preview of table data with optional filtering and sorting.\n\nArgs:\n    layer: Layer name (bronze, silver, gold)\n    table: Table name\n    preview_req: Preview request with filters and sort options",
        "operationId": "preview_table_api_data_preview__layer___table__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer"
            }
          },
          {
            "name": "table",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/query": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Execute Sql Query",
        "description": "Execute a SQL query against Delta tables.\n\nTables should be referenced as layer_table_name (e.g., bronze_accueillants, silver_geo).\nAll available Delta tables are automatically registered.\n\nArgs:\n    query_req: SQL query request with query string and optional limit",
        "operationId": "execute_sql_query_api_data_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_file_api_files_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_files_upload_post"
      },
      "CatalogResponse": {
        "properties": {
          "schemas": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/TableInfo"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Schemas"
          }
        },
        "type": "object",
        "required": [
          "schemas"
        ],
        "title": "CatalogResponse",
        "description": "Catalog response with all schemas and tables."
      },
...
      "FileUploadResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "destination": {
            "type": "string",
            "title": "Destination"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "destination",
          "size_bytes",
          "uploaded_at"
        ],
        "title": "FileUploadResponse",
        "description": "Response for file upload.",
        "example": {
          "destination": "raw/geo/communes_2024.csv",
          "filename": "communes_2024.csv",
          "size_bytes": 1048576,
          "uploaded_at": "2025-11-29T10:30:00Z"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewFilter": {
        "properties": {
          "column": {
            "type": "string",
            "title": "Column"
          },
          "operator": {
            "type": "string",
            "title": "Operator",
            "default": "="
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "column",
          "value"
        ],
        "title": "PreviewFilter",
        "description": "Filter specification for table preview."
      },
      "PreviewRequest": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 100
          },
          "filters": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PreviewFilter"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "sort_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort By"
          },
          "sort_order": {
            "type": "string",
            "title": "Sort Order",
            "default": "asc"
          }
        },
        "type": "object",
        "title": "PreviewRequest",
        "description": "Request for table preview."
      },
      "PreviewResponse": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "data": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "filtered_rows": {
            "type": "integer",
            "title": "Filtered Rows"
          },
          "preview_rows": {
            "type": "integer",
            "title": "Preview Rows"
          }
        },
        "type": "object",
        "required": [
          "columns",
          "data",
          "total_rows",
          "filtered_rows",
          "preview_rows"
        ],
        "title": "PreviewResponse",
        "description": "Table preview response."
      },
      "QueryRequest": {
        "properties": {
          "sql": {
            "type": "string",
            "title": "Sql"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 1000
          }
        },
        "type": "object",
        "required": [
          "sql"
        ],
        "title": "QueryRequest",
        "description": "Request for SQL query execution."
      },
      "QueryResponse": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "data": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "execution_time_ms": {
            "type": "number",
            "title": "Execution Time Ms"
          }
        },
        "type": "object",
        "required": [
          "columns",
          "data",
          "row_count",
          "execution_time_ms"
        ],
        "title": "QueryResponse",
        "description": "SQL query execution response."
      },
...
      "SchemaField": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "nullable": {
            "type": "boolean",
            "title": "Nullable"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "nullable"
        ],
        "title": "SchemaField",
        "description": "Schema field information."
      },
      "TableInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "version"
        ],
        "title": "TableInfo",
        "description": "Table information."
      },
      "TableSchema": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/SchemaField"
            },
            "type": "array",
            "title": "Fields"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "num_fields": {
            "type": "integer",
            "title": "Num Fields"
          }
        },
        "type": "object",
        "required": [
          "fields",
          "version",
          "row_count",
          "num_fields"
        ],
        "title": "TableSchema",
        "description": "Table schema information."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}